add installed checks

+++ packages/fonts.sh
@@ -2,9 +2,13 @@
 
 install() {
   if [[ "$IS_LINUX" == "true" ]]; then
-    execute mkdir -p ~/.fonts/JetBrainsMono
+    if [[ -d "$HOME/.fonts/JetBrainsMono" ]]; then
+      return
+    fi
+
+    execute mkdir -p $HOME/.fonts/JetBrainsMono
     execute eval "curl -LO https://github.com/ryanoasis/nerd-fonts/releases/latest/download/JetBrainsMono.tar.xz"
-    execute tar -xf JetBrainsMono.tar.xz -C ~/.fonts/JetBrainsMono
+    execute tar -xf JetBrainsMono.tar.xz -C $HOME/.fonts/JetBrainsMono
     execute rm JetBrainsMono.tar.xz
     execute fc-cache -f -v
   elif [[ "$IS_MACOS" == "true" ]]; then
@@ -15,6 +19,10 @@ install() {
 
 update() {
   if [[ "$IS_LINUX" == "true" ]]; then
-    install
+    execute mkdir -p $HOME/.fonts/JetBrainsMono
+    execute eval "curl -LO https://github.com/ryanoasis/nerd-fonts/releases/latest/download/JetBrainsMono.tar.xz"
+    execute tar -xf JetBrainsMono.tar.xz -C $HOME/.fonts/JetBrainsMono
+    execute rm JetBrainsMono.tar.xz
+    execute fc-cache -f -v
   fi
 }
